--- src/main/java/com/example/estsoft_udon_community/email/AuthCodeValidationException.java
@@ -1,7 +1,0 @@
-package com.example.estsoft_udon_community.email;
-
-public class AuthCodeValidationException extends RuntimeException {
-    public AuthCodeValidationException(String message) {
-        super(message);
-    }
-}
