+++ src/main/java/com/example/estsoft_udon_community/email/controller/EmailController.java
@@ -1,5 +1,8 @@
-package com.example.estsoft_udon_community.email;
+package com.example.estsoft_udon_community.email.controller;
 
+import com.example.estsoft_udon_community.email.AuthCodeGenerator;
+import com.example.estsoft_udon_community.email.service.AuthService;
+import com.example.estsoft_udon_community.email.service.EmailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,13 +13,13 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @RequiredArgsConstructor
 public class EmailController {
     private final EmailService emailService;
-    private final AuthCodeService authCodeService;
+    private final AuthService authService;
 
     @ResponseBody
     @PostMapping("/send-email")
     public String sendEmail(@RequestParam(required = false) String email) {
         String authCode = AuthCodeGenerator.generateCode(6); // 6자리 랜덤 숫자 생성
-        authCodeService.saveAuthCode(email, authCode);
+        authService.saveAuthCode(email, authCode);
         emailService.sendEmail(email, authCode);
         return "이메일 전송 완료!";
     }
