+++ src/main/java/com/example/estsoft_udon_community/email/service/EmailService.java
@@ -1,9 +1,8 @@
-package com.example.estsoft_udon_community.email;
+package com.example.estsoft_udon_community.email.service;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
